requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filebrowser-no-grappelli==4.0.1
django-tinymce4-lite==1.8.0
django-upgrade==1.12.0
jsmin==3.0.1
Pillow==9.3.0
six==1.16.0
sqlparse==0.4.3
tokenize-rt==5.0.0
tzdata==2022.6
